EOS-19912: Moved the key updation to config phase Signed-off-by: Vimal Raghwani <vimal.raghwani@seagate.com>
