+++ scripts/provisioning/configcmd.py
@@ -24,7 +24,8 @@ import errno
 import shutil
 from pathlib import Path
 from  ast import literal_eval
-
+from os import path
+from s3confstore.cortx_s3_confstore import S3CortxConfStore
 from setupcmd import SetupCmd, S3PROVError
 from cortx.utils.process import SimpleProcess
 from s3msgbus.cortx_s3_msgbus import S3CortxMsgBus
@@ -57,6 +58,10 @@ class ConfigCmd(SetupCmd):
     self.logger.info("validations completed")
 
     try:
+      self.logger.info("update motr max units per request started")
+      self.update_motr_max_units_per_request()
+      self.logger.info("update motr max units per request completed")
+
       self.logger.info('create auth jks password started')
       self.create_auth_jks_password()
       self.logger.info('create auth jks password completed')
@@ -241,3 +246,25 @@ class ConfigCmd(SetupCmd):
     else:
       self.logger.warning(f'warning of create_auth_jks_password.sh: {stderr}')
       self.logger.info(' Successfully set auth JKS keystore password.')
+
+  def update_motr_max_units_per_request(self):
+    """
+    update S3_MOTR_MAX_UNITS_PER_REQUEST in the s3config file based on VM/OVA/HW
+    S3_MOTR_MAX_UNITS_PER_REQUEST = 8 for VM/OVA
+    S3_MOTR_MAX_UNITS_PER_REQUEST = 32 for HW
+    """
+
+    # get the motr_max_units_per_request count from the config file
+    motr_max_units_per_request = self.get_confvalue(self.get_confkey('CONFIG>CONFSTORE_S3_MOTR_MAX_UNITS_PER_REQUEST'))
+    self.logger.info(f'motr_max_units_per_request: {motr_max_units_per_request}')
+
+    # update the S3_MOTR_MAX_UNITS_PER_REQUEST in s3config.yaml file
+    s3configfile = self.get_confkey('S3_CONFIG_FILE')
+    if path.isfile(f'{s3configfile}') == False:
+      self.logger.error(f'{s3configfile} file is not present')
+      raise S3PROVError(f'{s3configfile} file is not present')
+    else:
+      motr_max_units_per_request_key = 'S3_MOTR_CONFIG>S3_MOTR_MAX_UNITS_PER_REQUEST'
+      s3configfileconfstore = S3CortxConfStore(f'yaml://{s3configfile}', 'write_s3_motr_max_unit_idx')
+      s3configfileconfstore.set_config(motr_max_units_per_request_key, int(motr_max_units_per_request), True)
+      self.logger.info(f'Key {motr_max_units_per_request_key} updated successfully in {s3configfile}')
