+++ scripts/provisioning/templates/s3.config.tmpl.1-node.sample
@@ -34,6 +34,8 @@ cortx:
       sgiam:
         user: "sgiamadmin"
         secret: "gAAAAABgWFPz2w_t6IBVictZmgTCgBXn5V-oy2JUVtvZkJvHM5KaaZKXRwtz4YKYieIy8mVKF8PUaZ6WYJEttcsi2KDpLOk9Vg=="
+    s3:
+      motr_max_units_per_request : "8"
 cluster:
   3f670dd0-17cf-4ef3-9d8b-e1fb6a14c0f6:
     site:
