+++ scripts/provisioning/templates/s3.config.tmpl.3-node
@@ -50,6 +50,8 @@ cortx:
       sgiam:
         user: "sgiamadmin"
         secret: "TMPL_SGIAM_SECRET_KEY"
+    s3:
+      motr_max_units_per_request : "8"
 cluster:
   TMPL_CLUSTER_ID:
     site:
